generalize and use in vr
